requirements.txt
asgiref==3.8.1
boto3==1.37.37
botocore==1.37.37
certifi==2025.1.31
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
Django==5.2
django-allauth==65.7.0
django-daisy==1.0.22
django-storages==1.14.6
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
heroicons==2.11.0
idna==3.10
jmespath==1.0.1
msgpack==1.1.0
pillow==11.2.1
promise==2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
s3transfer==0.11.5
six==1.17.0
sqlparse==0.5.3
stripe==12.0.1
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
user-agents==2.2.0
zipstream-new==1.1.8
